+++ src/app/(SinglePage)/about/page.tsx
@@ -1,7 +1,8 @@
 import type { Metadata } from 'next';
-import { getMetadata, PAGE_CONFIGS } from '@/app/_metadata';
+import { getMetadata } from '@/app/_metadata';
 import { Template } from '@/components/Page/Single/Template';
 import { SITE_URL } from '@/constants';
+import { PAGE_CONFIGS } from '@/lib/page/config';
 
 export const metadata: Metadata = getMetadata({
   title: PAGE_CONFIGS.about.title,
