+++ src/app/not-found.tsx
@@ -1,5 +1,5 @@
-import { recentPosts } from '@/app/_lib/cachedRecentPosts';
 import { Content } from '@/components/Page/NotFound/Content';
+import { recentPosts } from '@/lib/post/list';
 
 export default function NotFound() {
   return <Content posts={recentPosts} />;
