+++ src/components/App/Search/engine/types.ts
@@ -1,0 +1,4 @@
+export type SearchDataPayload = {
+  searchIndex: SearchIndex;
+  searchData: SearchDataItem[];
+};
