--- src/components/Page/Single/types.ts
@@ -1,8 +1,0 @@
-/**
- * シングルページ設定の型定義
- */
-
-/**
- * 単一ページのスラグ文字列のユニオン型
- */
-export type PageSlug = 'about' | 'privacy';
